+++ packages/utils-for-node/lib/json.ts
@@ -1,6 +1,7 @@
 import fsExtra from 'fs-extra';
-import { packageJsonPath } from './path';
+import { join } from 'path';
+import { findNodeLibRootDir } from './path';
 const { readJson } = fsExtra;
 export async function getPackageJson() {
-  return await readJson(packageJsonPath);
+  return await readJson(join(findNodeLibRootDir(process.cwd()), 'package.json'));
 }