+++ packages/utils-for-node/lib/path/index.ts
@@ -1,0 +1,3 @@
+export * from "./path-resolver";
+export * from "./findUp";
+export * from "./findNodeLibRootDir";