+++ tools/version/lib/getter/Getter.ts
@@ -1,3 +1,4 @@
 export interface Getter {
   get(): Promise<string | number> | string | number;
 }
+