+++ tools/create-tools/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tikkhun/create
 
+## 0.2.10
+
+### Patch Changes
+
+- feat: findUp file 以防在子目录上运行没有效果
+- Updated dependencies
+  - @tikkhun/utils@1.0.17
+
 ## 0.2.9
 
 ### Patch Changes