+++ tools/release-plugins/info/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tikkhun/info
 
+## 0.0.5
+
+### Patch Changes
+
+- feat: findUp file 以防在子目录上运行没有效果
+- Updated dependencies
+  - @tikkhun/utils@1.0.17
+
 ## 0.0.4
 
 ### Patch Changes