+++ tools/release-plugins/input-move/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tikkhun/release-input-move
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @tikkhun/info@0.0.5
+
 ## 0.0.2
 
 ### Patch Changes